\name{countries}
\alias{countries}
\docType{data}
\title{
Names of the countries in the Human Mortality Database
}
\description{
Names of the countries for which data are available in this package. 
}
\usage{data("countries")}
\format{
  vector of character strings
}
\examples{
data(countries)
print(countries)
}

