\name{fin.lca}
\alias{fin.lca}
\docType{data}
\title{
Lee-Carter model fit for Finnish data
}
\description{Lee-Carter model fit obtained by function "lca" of package "demography". The fit
is based on Finnish mortality data for ages from 0 to 100 and years from 1950 to 2015. 
}
\usage{data("fin.lca")}
\format{
object of class "lca"  
}
\details{
Both sexes were included in the input mortality data. 
}
\examples{
data(fin.lca)
plot(fin.lca) 
}

