\name{RcmdrPlugin.RiskDemo-package}
\alias{RcmdrPlugin.RiskDemo-package}
\alias{RcmdrPlugin.RiskDemo}
\docType{package}
\title{
R Commander Plug-in for Risk Demonstration
}
\description{
R Commander plug-in to demonstrate various actuarial and financial risks. It includes valuation of bonds and stocks, portfolio optimization, classical ruin theory and demography. 
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.RiskDemo\cr
Type: \tab Package\cr
Version: \tab 1.9\cr
Date: \tab 2017-08-09\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Arto Luoma

Maintainer: Arto Luoma <arto.luoma@wippies.com>
}

\keyword{ package }


